Separated the sound engine from the android app
The code still lives in this repo, it just now is handled as a separate
library, so I can develop it independently of the UI in the android app.  I
also wrote a basic Linux interface to the sound engine, so I can do development
of the engine more easily.  That code probably shouldn't live in the library,
but it's ok for now.