Refactoring and work on music rendering
I moved the desktop-specific code to its own set of packages, so it'll be easy
to move to its own application later, and I put together a framework for music
rendering.  Most of the work should now be done in StaffDrawer, now that I have
the other classes well defined and working.